+++ cointrade/app/datasrc/DataSrcTrait.scala
@@ -4,5 +4,5 @@ import scala.concurrent._
 
 trait DataSrcTrait {
  
-  def priceInTimeRange(coin:String,startTime : Long, endTime : Long) :Future[List[Int]]
+  def priceInTimeRange(coin:String,startTime : Long, endTime : Long) :Future[List[Long]]
 }
